requirements.txt
groq>=0.11.0
instructor>=1.7.0
pydantic>=2.10.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
streamlit>=1.40.0
python-dotenv>=1.0.0
requests>=2.32.0
